+++ NEWS.md
@@ -4,4 +4,27 @@
 
 * Added support for edge labels with significant coefficients
 
+# lavaanPlot 0.4.0
+
+* Added support for latent variable loading labels
+
+* Added support for covariance edges
+
+* Corrected the direction of arrows for latent variable loadings
+
+# lavaanPlot 0.5.0
+
+* Added support for covariance edge labels
+
+* Added support for significance stars
+
+# lavaanPlot 0.5.1
+
+* Fixed bug caused by . in variable names
+
+* simplified the functions using the ... argument
+
+* fixed significance stars to work with standardized and non-standardized coefficients
+
+* Changed significance stars so you can choose which parameters to use them with